requirements.txt
fastapi
click
GitPython
uvicorn[standard]
pydantic
python-multipart
